requirements.txt
numpy
pickle-mixin
streamlit
streamlit-option-menu
scikit-learn
protobuf==3.20.1
Pillow
